src_async/tools_com/com_Material/com_camera.py
"""
Module com_camera - Fonctions de présence, fonctionnement et utilisation de la caméra
"""

import logging

logger = logging.getLogger(__name__)

def check_camera_presence():
    """
    Vérifie la présence de la caméra.

    :return: True si la caméra est présente, False sinon.
    :rtype: bool
    """
    try:
        # Vérification de la présence de la caméra ici
        presence = True  # Exemple de réponse (à adapter)
        if presence:
            logger.info("Camera is present.")
        else:
            logger.info("Camera is not present.")
        return presence
    except Exception:
        logger.error("Failed to check camera presence.")
        return False

def start_camera():
    """
    Démarre la caméra pour la capture d'images ou de vidéos.

    :return: True si la caméra a démarré avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Démarrage de la caméra ici
        started = True  # Exemple de réponse (à adapter)
        if started:
            logger.info("Camera started.")
        else:
            logger.error("Failed to start camera.")
        return started
    except Exception:
        logger.error("Failed to start camera.")
        return False

def capture_image():
    """
    Capture une image à partir de la caméra.

    :return: True si l'image a été capturée avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Capture d'une image à partir de la caméra ici
        image_captured = True  # Exemple de réponse (à adapter)
        if image_captured:
            logger.info("Image captured.")
        else:
            logger.error("Failed to capture image.")
        return image_captured
    except Exception:
        logger.error("Failed to capture image.")
        return False

def record_video(duration):
    """
    Enregistre une vidéo à partir de la caméra pendant la durée spécifiée.

    :param duration: Durée de l'enregistrement vidéo en secondes.
    :type duration: int
    :return: True si la vidéo a été enregistrée avec succès, False sinon.
    :rtype: bool
    """
    try:
        # Enregistrement d'une vidéo à partir de la caméra ici
        video_recorded = True  # Exemple de réponse (à adapter)
        if video_recorded:
            logger.info("Video recorded.")
        else:
            logger.error("Failed to record video.")
        return video_recorded
    except Exception:
        logger.error("Failed to record video.")
        return False
